Disable trimming when zoomed
This is a major revert of functionaility. We decided that trimming
while zoomed in was not worth the zoom accuracy bugs we kept hitting.
Instead, we are disabling trimming when zoomed in. This means
reverting a swath of patches.
We still kept the logic that determines the zoom region based on
the data points inside it. This logic is slightly changed to use
the timeline time of the data points rather than the pixel position.
This commit also includes the other improvements added during the zoom
process:
  - Zoom extent covers the rulers
  - Last lane of timeline doesn't collide with edge if tracks disabled
  - y/m/d notation for ruler
  - Single click zooming